more plugins refactor
